pyproject.toml
[tool.poetry]
name = "kg-valute-parser"
version = "0.1.0"
description = "Parse currency course for kg"
authors = ["Artyom Kropp <artyomkropp@gmail.com>"]
readme = "README.md"
packages = [{include = "kg_valute_parser"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "3.0.0b5"
python-dotenv = "^0.21.0"
pandas = "^1.5.1"
requests = "^2.28.1"
lxml = "^4.9.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
wemake-python-styleguide = "^0.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
